+++ apps/web/app/(use-page-wrapper)/apps/(homepage)/page.tsx
@@ -1,10 +1,12 @@
-import { withAppDirSsr } from "app/WithAppDirSsr";
-import type { PageProps } from "app/_types";
 import { _generateMetadata } from "app/_utils";
 import { cookies, headers } from "next/headers";
 
-import { getServerSideProps } from "@lib/apps/getServerSideProps";
-import { buildLegacyCtx } from "@lib/buildLegacyCtx";
+import { getAppRegistry, getAppRegistryWithCredentials } from "@calcom/app-store/_appRegistry";
+import { getServerSession } from "@calcom/features/auth/lib/getServerSession";
+import { UserRepository } from "@calcom/lib/server/repository/user";
+import type { AppCategories } from "@calcom/prisma/enums";
+
+import { buildLegacyRequest } from "@lib/buildLegacyCtx";
 
 import AppsPage from "~/apps/apps-view";
 
@@ -18,13 +20,43 @@ export const generateMetadata = async () => {
   );
 };
 
-const getData = withAppDirSsr(getServerSideProps);
+const ServerPage = async () => {
+  const req = buildLegacyRequest(await headers(), await cookies());
+  const session = await getServerSession({ req });
+  let appStore, userAdminTeamsIds: number[];
+  if (session?.user?.id) {
+    const userAdminTeams = await UserRepository.getUserAdminTeams(session.user.id);
+    userAdminTeamsIds = userAdminTeams?.teams?.map(({ team }) => team.id) ?? [];
+    appStore = await getAppRegistryWithCredentials(session.user.id, userAdminTeamsIds);
+  } else {
+    appStore = await getAppRegistry();
+    userAdminTeamsIds = [];
+  }
 
-const ServerPage = async ({ params, searchParams }: PageProps) => {
-  const context = buildLegacyCtx(await headers(), await cookies(), await params, await searchParams);
-  const props = await getData(context);
+  const categoryQuery = appStore.map(({ categories }) => ({
+    categories: categories || [],
+  }));
+  const categories = categoryQuery.reduce((c, app) => {
+    for (const category of app.categories) {
+      c[category] = c[category] ? c[category] + 1 : 1;
+    }
+    return c;
+  }, {} as Record<string, number>);
 
-  return <AppsPage {...props} />;
+  const props = {
+    categories: Object.entries(categories)
+      .map(([name, count]): { name: AppCategories; count: number } => ({
+        name: name as AppCategories,
+        count,
+      }))
+      .sort(function (a, b) {
+        return b.count - a.count;
+      }),
+    appStore,
+    userAdminTeams: userAdminTeamsIds,
+  };
+
+  return <AppsPage {...props} isAdmin={session?.user?.role === "ADMIN"} />;
 };
 
 export default ServerPage;
